Move gestures to own package
