+++ packages/gestures/__tests__/ios/gestures.ios.test.js
@@ -1,0 +1,14 @@
+import tests from "../gestures.native.test";
+
+jest.mock("react-native", () => {
+  const reactNative = require.requireActual("react-native");
+  reactNative.Platform.OS = "ios";
+  jest
+    .spyOn(reactNative.Platform, "select")
+    .mockImplementation(obj => obj.ios || obj.default);
+  return reactNative;
+});
+
+describe("Gesture tests", () => {
+  tests();
+});
