+++ packages/gestures/package.json
@@ -1,0 +1,63 @@
+{
+  "name": "@times-components/gestures",
+  "version": "1.0.0",
+  "main": "gestures",
+  "scripts": {
+    "test:android": "jest --config='./__tests__/android/jest.config.js'",
+    "test:android:debug":
+      "node --inspect --inspect-brk node_modules/.bin/jest --config='./__tests__/android/jest.config.js'",
+    "test:ios": "jest --config='./__tests__/ios/jest.config.js'",
+    "test": "yarn test:android --ci --bail --coverage && yarn test:ios --ci --bail --coverage",
+    "test:watch": "jest --projects */**/jest.config.js --watchAll",
+    "fmt": "prettier --write '**/*.*'",
+    "prettier:diff": "prettier --list-different '**/*.*'",
+    "lint": "eslint . && yarn prettier:diff",
+    "dextrose-stories": "dextrose-stories",
+    "dextrose-clean": "dextrose-clean"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/newsuk/times-components.git"
+  },
+  "keywords": [
+    "react-native-web",
+    "react",
+    "native",
+    "web",
+    "gestures",
+    "component"
+  ],
+  "license": "BSD-3-Clause",
+  "bugs": {
+    "url": "https://github.com/newsuk/times-components/issues"
+  },
+  "homepage": "https://github.com/newsuk/times-components#readme",
+  "devDependencies": {
+    "@storybook/react-native": "3.3.9",
+    "@times-components/eslint-config-thetimes": "0.1.2",
+    "dextrose": "1.3.2",
+    "enzyme": "3.3.0",
+    "enzyme-to-json": "3.3.0",
+    "eslint": "4.9.0",
+    "jest": "21.2.1",
+    "prettier": "1.8.2",
+    "react": "16.2.0",
+    "react-native": "0.50.1",
+    "react-native-web": "0.3.2",
+    "react-test-renderer": "16.2.0"
+  },
+  "dependencies": {
+    "@times-components/jest-configurator": "0.0.21",
+    "prop-types": "15.6.0"
+  },
+  "peerDependencies": {
+    "babel-preset-env": "1.6.1",
+    "babel-preset-stage-2": "6.24.1",
+    "react": ">=16",
+    "react-native": ">=0.50",
+    "react-native-web": ">=0.3"
+  },
+  "publishConfig": {
+    "access": "public"
+  }
+}
